api_doc_agent/requirements.txt
phidata
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain-community==0.0.38
langchain-text-splitters==0.0.1
chromadb==0.5.0
sentence-transformers==2.7.0
groq==0.8.0
aiofiles==23.2.1 # For async file operations if needed, though not directly used here